+++ src/bidiMapper/domains/events/EventManager.ts
@@ -21,8 +21,8 @@ import {
   InvalidArgumentException,
 } from '../../../protocol/protocol.js';
 import {DefaultMap} from '../../../utils/DefaultMap.js';
-import {Buffer} from '../../../utils/buffer.js';
-import {IdWrapper} from '../../../utils/idWrapper.js';
+import {Buffer} from '../../../utils/Buffer.js';
+import {IdWrapper} from '../../../utils/IdWrapper.js';
 import type {Result} from '../../../utils/result.js';
 import type {BidiServer} from '../../BidiServer.js';
 import {OutgoingMessage} from '../../OutgoingMessage.js';
