+++ src/bidiMapper/domains/network/NetworkRequest.ts
@@ -22,7 +22,7 @@
  */
 import type Protocol from 'devtools-protocol';
 
-import {Deferred} from '../../../utils/deferred.js';
+import {Deferred} from '../../../utils/Deferred.js';
 import type {EventManager} from '../events/EventManager.js';
 import {
   type Network,
