+++ src/bidiServer/MapperServer.ts
@@ -19,10 +19,10 @@ import type Protocol from 'devtools-protocol';
 import WebSocket from 'ws';
 import debug, {type Debugger} from 'debug';
 
-import {CdpConnection} from '../cdp/cdpConnection.js';
-import type {CdpClient} from '../cdp/cdpClient.js';
+import {CdpConnection} from '../cdp/CdpConnection.js';
+import type {CdpClient} from '../cdp/CdpClient.js';
 import type {LogType, LogPrefix} from '../utils/log.js';
-import {WebSocketTransport} from '../utils/websocketTransport.js';
+import {WebSocketTransport} from '../utils/WebsocketTransport.js';
 
 const debugInternal = debug('bidi:mapper:internal');
 const debugInfo = debug('bidi:mapper:info');
